requirements.txt
PyQt6==6.7.1
Pillow>=10.0.0
